Przeniesiono komponenty z OrderAWalk, poprawki css
