+++ frontend/react-frontend/src/Flows/ClientFlow/ClientProfile/ClientProfile.css
@@ -12,6 +12,8 @@
   -webkit-box-align: center;
       -ms-flex-align: center;
           align-items: center;
+  margin-top: 30px;
+  margin-bottom: 30px;
 }
 
 @media screen and (max-width: 660px) {
