+++ frontend/react-frontend/src/Flows/ClientFlow/ClientProfile/ClientProfile.scss
@@ -21,6 +21,8 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  margin-top: 30px;
+  margin-bottom: 30px;
 
   h1 {
     @include text($fontSize: 36px);
