+++ frontend/react-frontend/src/Flows/TrainerFlow/TrainerProfile/TrainerProfile.scss
@@ -22,6 +22,8 @@
   display: flex;
   flex-direction: column;
   align-items: center;
+  margin-top: 30px;
+  margin-bottom: 30px;
 
   h1 {
     @include text($fontSize: 36px);
